Implement Stack using Linked list
This program implements a stack data structure using a linked list. The program prompts the user to enter the maximum size of the stack and creates a stack object with the given maximum size. It then performs various operations on the stack, such as pushing and popping elements, displaying the elements in the stack, and checking if the stack is empty or full.
The program also uses the <chrono> library to measure the time taken to perform these operations and outputs the duration in microseconds. The time measurement can be used to compare the efficiency of different implementations of the stack data structure.